+++ changelog.md
@@ -1,3 +1,10 @@
+### v1.2.0
+
+* Add `owner` option.
+* Add `group` option.
+* Add tests.
+* Add `create` method.
+
 ### v1.1.0
 
 * Add `SYNCFILE` environment variable.